// frontends/main/src/app/core/model/field-model-dto.ts
/**
 * Rodano API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { FieldModelType } from './field-model-type';
import { OperandType } from './operand-type';
import { PossibleValueDTO } from './possible-value-dto';


export interface FieldModelDTO { 
    /**
     * Field type
     */
    type: FieldModelType;
    /**
     * Field value type for the operands
     */
    dataType: OperandType;
    datasetModelId: string;
    id: string;
    shortname: { [key: string]: string; };
    longname: { [key: string]: string; };
    description?: { [key: string]: string; };
    /**
     * Possible values of the field, if the field type is a multiple
     */
    possibleValues: Array<PossibleValueDTO>;
    dictionary?: string;
    readOnly: boolean;
    dynamic: boolean;
    order?: number;
    /**
     * Can the field be found in search
     */
    searchable: boolean;
    /**
     * Is the field model included in the exports
     */
    exportable: boolean;
    required: boolean;
    maxLength?: number;
    inlineHelp?: string;
    advancedHelp?: { [key: string]: string; };
    minValue?: number;
    maxValue?: number;
    minYear?: number;
    maxYear?: number;
    maxIntegerDigits?: number;
    maxDecimalDigits?: number;
    allowDateInFuture: boolean;
    withDays: boolean;
    withMonths: boolean;
    withYears: boolean;
    withHours: boolean;
    withMinutes: boolean;
    withSeconds: boolean;
    yearsMandatory: boolean;
    monthsMandatory: boolean;
    daysMandatory: boolean;
    hoursMandatory: boolean;
    minutesMandatory: boolean;
    secondsMandatory: boolean;
}
export namespace FieldModelDTO {
}